refactor: update hook for todo list, update todo list
