+++ frontend/src/components/TodoTaskList/TodoTaskListFooter/index.styled.tsx
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const StyledTodoTaskListFooter = styled.div`
+  font-weight: 600;
+  font-size: 14px;
+  text-transform: uppercase;
+  text-align: center;
+  cursor: pointer;
+`;
