+++ frontend/src/components/TodoTaskList/TodoTaskListFooter/index.tsx
@@ -1,8 +1,13 @@
 import { TodoTaskListFooterProps } from './types';
+import { StyledTodoTaskListFooter } from './index.styled';
 
 export const TodoTaskListFooter = ({
   completedTaskVisible,
   handleCompleteTasks,
 }: TodoTaskListFooterProps) => {
-  return <div></div>;
+  return (
+    <StyledTodoTaskListFooter onClick={handleCompleteTasks}>
+      <span>{completedTaskVisible ? 'Hide complete tasks' : 'Show complete tasks'}</span>
+    </StyledTodoTaskListFooter>
+  );
 };
