+++ frontend/src/components/TodoTaskList/index.styled.tsx
@@ -1,2 +1,7 @@
+import { List } from 'antd';
 import styled from 'styled-components';
-export {};
+
+// ? What am I doing wrong?
+export const StyledList = styled(List)`
+  background-color: #fff;
+`;
