+++ frontend/src/views/TodoList/TodoList.tsx
@@ -1,5 +1,7 @@
 import { useEffect } from 'react';
 import { usePageTitle } from 'providers/PageTitleProvider';
+import { TodoTaskList } from 'components/TodoTaskList';
+import { ListMock } from 'mock-data/TodoTaskList/TodotaskListMock';
 
 export const TodoList = () => {
   const { setTitle } = usePageTitle();
@@ -8,5 +10,9 @@ export const TodoList = () => {
     setTitle('To Do List');
   }, []);
 
-  return <>To Do List</>;
+  return (
+    <>
+      <TodoTaskList todoTasks={ListMock.todoTasks} />
+    </>
+  );
 };
